- Fixed: custom tab now only displayed when there are html files in the template folder, and the option display custom tab is enabled - udevice json call accepts now updates without nvalue/svalue, still recommended to use both

+++ trunk/main/mainworker.cpp
@@ -10664,7 +10664,7 @@ void MainWorker::HeartbeatCheck()
 	}
 }
 
-bool MainWorker::UpdateDevice(const int idx, const int nValue, const std::string &sValue, const int signallevel, const int batterylevel)
+bool MainWorker::UpdateDevice(const unsigned long long idx, const int nValue, const std::string &sValue, const int signallevel, const int batterylevel)
 {
 	std::stringstream szQuery;
 	std::vector<std::vector<std::string> > result;
